Move api connection related code to pkg. (#336)
* Move api connection related code to pkg.

* Fix module name.
