+++ cmd/armadactl/cmd/submit.go
@@ -66,7 +66,7 @@ var submitCmd = &cobra.Command{
 
 		requests := client.CreateChunkedSubmitRequests(submitFile.Queue, submitFile.JobSetId, submitFile.Jobs)
 
-		util.WithConnection(apiConnectionDetails, func(conn *grpc.ClientConn) {
+		client.WithConnection(apiConnectionDetails, func(conn *grpc.ClientConn) {
 			submissionClient := api.NewSubmitClient(conn)
 			for _, request := range requests {
 				response, e := client.SubmitJobs(submissionClient, request)
