+++ pkg/client/auth/kerberos/common.go
