+++ pkg/client/auth/kerberos/spnego_unix.go
