+++ pkg/client/auth/kerberos/spnego_windows.go
